# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 2.0.0
  description: TransportPro - Transportation Business Management API
servers:
  - url: /api
...
paths:
...
  /subcontractors/{id}:
    get:
      operationId: getSubcontractor
      tags: [subcontractors]
      summary: Get subcontractor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubcontractorDetail"
    put:
      operationId: updateSubcontractor
      tags: [subcontractors]
      summary: Update subcontractor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubcontractorInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcontractor"
    delete:
      operationId: deleteSubcontractor
      tags: [subcontractors]
      summary: Delete subcontractor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /subcontractor-transactions/{id}:
    put:
      operationId: updateSubcontractorTransaction
      tags: [subcontractors]
      summary: Update subcontractor transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubcontractorTransactionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubcontractorTransaction"
    delete:
      operationId: deleteSubcontractorTransaction
      tags: [subcontractors]
      summary: Delete subcontractor transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trucks:
    get:
      operationId: getTrucks
      tags: [trucks]
      summary: List trucks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Truck"
    post:
      operationId: createTruck
      tags: [trucks]
      summary: Create truck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TruckInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
  /trucks/{id}:
    get:
      operationId: getTruck
      tags: [trucks]
      summary: Get truck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
    put:
      operationId: updateTruck
      tags: [trucks]
      summary: Update truck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TruckInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
    delete:
      operationId: deleteTruck
      tags: [trucks]
      summary: Delete truck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip detail with financials
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
    put:
      operationId: updateTrip
      tags: [trips]
      summary: Update trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Remove trip from batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /trip-expenses/{id}:
    put:
      operationId: updateTripExpense
      tags: [expenses]
      summary: Update trip expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripExpenseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripExpense"
    delete:
      operationId: deleteTripExpense
      tags: [expenses]
      summary: Delete trip expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Subcontractor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        commissionRate:
          type: number
        notes:
          type: string
          nullable: true
        balance:
          type: number
          description: Outstanding balance (positive = we owe them)
        truckCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, commissionRate, balance, truckCount, createdAt]
    SubcontractorDetail:
      allOf:
        - $ref: "#/components/schemas/Subcontractor"
        - type: object
          properties:
            transactions:
              type: array
              items:
                $ref: "#/components/schemas/SubcontractorTransaction"
            trucks:
              type: array
              items:
                $ref: "#/components/schemas/Truck"
          required: [transactions, trucks]
    SubcontractorInput:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        commissionRate:
          type: number
        notes:
          type: string
          nullable: true
      required: [name, commissionRate]
    SubcontractorTransaction:
      type: object
      properties:
        id:
          type: integer
        subcontractorId:
          type: integer
        type:
          type: string
          enum: [net_payable, advance_given, payment_made, driver_salary, adjustment]
        amount:
          type: number
        tripId:
          type: integer
          nullable: true
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        transactionDate:
          type: string
        createdAt:
          type: string
      required: [id, subcontractorId, type, amount, transactionDate, createdAt]
    SubcontractorTransactionInput:
      type: object
      properties:
        type:
          type: string
          enum: [net_payable, advance_given, payment_made, driver_salary, adjustment]
        amount:
          type: number
        tripId:
          type: integer
          nullable: true
        driverId:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        transactionDate:
          type: string
      required: [type, amount, transactionDate]
...
    Truck:
      type: object
      properties:
        id:
          type: integer
        plateNumber:
          type: string
        trailerPlate:
          type: string
          nullable: true
        subcontractorId:
          type: integer
        subcontractorName:
          type: string
        status:
          type: string
          enum: [available, on_trip, maintenance, idle]
        notes:
          type: string
          nullable: true
        assignedDriverName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, plateNumber, subcontractorId, subcontractorName, status, createdAt]
    TruckInput:
      type: object
      properties:
        plateNumber:
          type: string
        trailerPlate:
          type: string
          nullable: true
        subcontractorId:
          type: integer
        status:
          type: string
          enum: [available, on_trip, maintenance, idle]
        notes:
          type: string
          nullable: true
      required: [plateNumber, subcontractorId, status]
...
    Trip:
      type: object
      properties:
        id:
          type: integer
        batchId:
          type: integer
        batchName:
          type: string
          nullable: true
        truckId:
          type: integer
        truckPlate:
          type: string
        trailerPlate:
          type: string
          nullable: true
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        subcontractorId:
          type: integer
        subcontractorName:
          type: string
        product:
          type: string
          enum: [AGO, PMS]
        capacity:
          type: number
        status:
          type: string
          enum: [nominated, amended_out, cancelled, loading, loaded, in_transit, at_zambia_entry, at_drc_entry, delivered]
        loadedQty:
          type: number
          nullable: true
        deliveredQty:
          type: number
          nullable: true
        mileageStart:
          type: number
          nullable: true
        mileageEnd:
          type: number
          nullable: true
        fuel1:
          type: number
          nullable: true
        fuel2:
          type: number
          nullable: true
        fuel3:
          type: number
          nullable: true
        cancellationReason:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, batchId, truckId, truckPlate, subcontractorId, subcontractorName, product, capacity, status, createdAt]
    TripDetail:
      allOf:
        - $ref: "#/components/schemas/Trip"
        - type: object
          properties:
            expenses:
              type: array
              items:
                $ref: "#/components/schemas/TripExpense"
            clearances:
              type: array
              items:
                $ref: "#/components/schemas/Clearance"
            amendments:
              type: array
              items:
                $ref: "#/components/schemas/TripAmendment"
            deliveryNote:
              $ref: "#/components/schemas/DeliveryNote"
              nullable: true
            financials:
              $ref: "#/components/schemas/TripFinancials"
          required: [expenses, clearances, amendments, financials]
    TripFinancials:
      type: object
      properties:
        grossRevenue:
          type: number
          nullable: true
        commission:
          type: number
          nullable: true
        shortQty:
          type: number
          nullable: true
        allowancePct:
          type: number
          nullable: true
        allowanceQty:
          type: number
          nullable: true
        chargeableShort:
          type: number
          nullable: true
        shortCharge:
          type: number
          nullable: true
        tripExpensesTotal:
          type: number
        driverSalaryAllocation:
          type: number
        netPayable:
          type: number
          nullable: true
      required: [tripExpensesTotal, driverSalaryAllocation]
    TripUpdateInput:
      type: object
      properties:
        driverId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [nominated, amended_out, cancelled, loading, loaded, in_transit, at_zambia_entry, at_drc_entry, delivered]
        loadedQty:
          type: number
          nullable: true
        deliveredQty:
          type: number
          nullable: true
        mileageStart:
          type: number
          nullable: true
        mileageEnd:
          type: number
          nullable: true
        fuel1:
          type: number
          nullable: true
        fuel2:
          type: number
          nullable: true
        fuel3:
          type: number
          nullable: true
        cancellationReason:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    TripAmendment:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        amendmentType:
          type: string
        oldTruckId:
          type: integer
          nullable: true
        newTruckId:
          type: integer
          nullable: true
        oldDriverId:
          type: integer
          nullable: true
        newDriverId:
          type: integer
          nullable: true
        reason:
          type: string
        amendedAt:
          type: string
      required: [id, tripId, amendmentType, reason, amendedAt]
    TripExpense:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        costType:
          type: string
        description:
          type: string
          nullable: true
        amount:
          type: number
        currency:
          type: string
        createdAt:
          type: string
      required: [id, tripId, costType, amount, currency, createdAt]
    TripExpenseInput:
      type: object
      properties:
        costType:
          type: string
        description:
          type: string
          nullable: true
        amount:
          type: number
        currency:
          type: string
          default: USD
      required: [costType, amount]
    Clearance:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        checkpoint:
          type: string
          enum: [zambia_entry, drc_entry]
        documentType:
          type: string
        documentNumber:
          type: string
          nullable: true
        status:
          type: string
          enum: [requested, pending, approved, rejected]
        requestedAt:
          type: string
          nullable: true
        approvedAt:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, tripId, checkpoint, documentType, status, createdAt]
...
    DeliveryNote:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        content:
          type: string
        attachmentUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, tripId, content, createdAt]
...